CHROMIUM: config: Turn off QCOM_SMD_RPM for qcom config
None of the devices building from the Qualcomm config in this kernel
tree need "RPM".  They all use RPMH.  Let's turn it off until there is
a reason to turn it on.  It's not like we have all the other configs
that would be needed for boards using RPM anyway.

This is:

  echo CONFIG_QCOM_SMD_RPM=n >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
  ./chromeos/scripts/kernelconfig olddefconfig

...atop an already normalized config.

BUG=b:143638832
TEST=Build and boot

Change-Id: I524fd8a25f9aff8eedd9f28a01e41cd12389c7b0
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2015007
Reviewed-by: Stephen Boyd <swboyd@chromium.org>