FROMGIT: ASoC: SOF: Intel: cnl: Implement feature to support DSP D0i3 in S0
This patch implements support for DSP D0i3 when the system
is in S0. The basic idea is to schedule a delayed work after
every successful IPC TX that checks if there are only
D0I3-compatible streams active and if so transition
the DSP to D0I3.

With the introduction of DSP D0I3 in S0, we need to
ensure that the DSP is in D0I0 before sending any new
IPCs. The exception for this would be the
compact IPCs that are used to set the DSP in
D0I3/D0I0 states.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20200129220726.31792-9-pierre-louis.bossart@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 63e51fd33fef04b634a0c32ae491ab16a19cb17c
https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
for-next)

Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:149102665
TEST=Audio playback via HDMI/DP
TEST=Suspend to S0iX with audio playback with right coreboot

Change-Id: I505b7c233ec078922c97734e7a595a1740274b41
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2070512
Reviewed-by: Sean Paul <seanpaul@chromium.org>