FROMGIT: media: venus: vdec: Use pmruntime autosuspend
Implement pmruntime autosuspend in video decoder. This will
allow to save power while the userspace is inactive for some
reasonable period of time. Here we power-off venus core clocks
and power domain and don't touch vcodec because it is under
hardware control. The later decision is made to simplify the
code and avoid a mess in the power management code.

Signed-off-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 63342afea65e9efa06889775c27001745e157770
 git://linuxtv.org/svarbanov/media_tree.git venus-for-next-v5.8-part2)

BUG=b:149955289
TEST=video decoding still works on trogdor:
  watch -n .1 "cat /sys/kernel/debug/clk/*venus*/clk_rate"
  play youtube video
    => clock rates go up

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I40b09aad0bb36d127b95347c26b3514fc0ada101
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2172662
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>