CHROMIUM: WAR: spi: qup: fix spi_qup_set_cs handler
There is a possibility of spi_qup_set_cs() causing synchronus
external abort when spi device is in suspend state, as it will
be an unclocked access. So resume the spi device if it is in
suspend state when spi_qup_set_cs() functionality is invoked.

BUG=b:147624552
TEST=Compile,bootup and successful probe of spi devices

Change-Id: Ie435ce989eabfa1acb309de9d28797cc956333f8
Signed-off-by: Vijay Navnath Kamble <vkambl@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2036461
Reviewed-by: Yuji Sasaki <sasakiy@chromium.org>
Reviewed-by: SANTHOSH JANARDHANA HASSAN <sahassan@google.com>
Commit-Queue: Yuji Sasaki <sasakiy@chromium.org>
Tested-by: Yuji Sasaki <sasakiy@chromium.org>