FROMLIST: platform/x86: intel_scu_ipc: Log more information if SCU IPC command fails
Currently we only log an error if the command times out which makes it
hard to figure out the failing command. This changes the driver to log
command and subcommand with the error code which should make debugging
easier. This also allows us to simplify the callers as they don't need
to log these errors themselves.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
(am from https://patchwork.kernel.org/patch/11386447/)

BUG=b:147460168
TEST=Thunderbolt dock is detected and display behind dock connected to HoHo comes up.

Change-Id: I4ab60cd96364f3d361064e7b4133ac992802f8c0
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>