FROMGIT: ASoC: rt5682: Add CCF usage for providing I2S clks
There is a need to use RT5682 as DAI clock master for other codecs
within a platform, which means that the DAI clocks are required to
remain, regardless of whether the RT5682 is actually running
playback/capture.

The RT5682 CCF basic functions are implemented almost by the existing
internal functions and asoc apis. It needs a clk provider (rt5682 mclk)
to generate the bclk and wclk outputs.

The RT5682 CCF supports and restricts as below:
1. Fmt of DAI-AIF1 must be configured to master before using CCF.
2. Only accept a 48MHz clk as the clk provider.
3. Only provide a 48kHz wclk and a set of multiples of wclk as bclk.

There are some temporary limitations in this patch until a better
implementation.

Signed-off-by: Derek Fang <derek.fang@realtek.com>
Link: https://lore.kernel.org/r/1582033912-6841-1-git-send-email-derek.fang@realtek.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit ebbfabc16d23dfd20eecd4b6e68212fec37ae7c6
 https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:150663660
TEST=Capture and playback tested on Trembyle

Signed-off-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Change-Id: Ia912d9c61007dbc3f9299de1773f297a59507bd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2100194
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>