FROMGIT: coresight: Pass coresight_device for coresight_release_platform_data
As we prepare to expose the links between the devices in
sysfs, pass the coresight_device instance to the
coresight_release_platform_data in order to free up the connections
when the device is removed.

No functional changes as such in this patch.

Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Mike Leach <mike.leach@linaro.org>
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
(cherry picked from commit c8d72a60b92187d563e0ca07ee3378f8706b21a4
 git://git.linaro.org/kernel/coresight.git next)

BUG=b:149851136
TEST=See /sys/bus/coresight has devices on Trogdor

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I9df205012c9b273454e5a3e164da21b0c94d02b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2161815
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>