UPSTREAM: drm/i915/gen11: Moving WAs to rcs_engine_wa_init()
This are register of render engine, so after a render reset those
would return to the default value and init_clock_gating() is not
called for single engine reset.
So here moving it rcs_engine_wa_init() that will guarantee that this
WAs will not be lost.

Cc: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200302231421.224322-1-jose.souza@intel.com
(cherry picked from commit 3551ff92874405674b58ac0f5742069a5ddfc462)

BUG=b:175077993
TEST=Tested on eve-kernelnext

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I47d43dfae3ac85f7d65f6a09c7d2fe9057aa63eb