HACK: CHROMIUM: vidc: vdec: changes to enable UBWC on sc7180
Driver chagnes to enable UBWC on sc7180

There currently isn't any way of using modifiers to tell
v4l2 that a buffer is expecting UBWC video and that the
decoder should use UBWC when decoding.

The long term approach is to use ioctls extensions that
have been purposed on the ml.
https://lore.kernel.org/linux-media/20191008091119.7294-1-boris.brezillon@collabora.com/

BUG=b:149525848
TEST=build kernel for trogdor

Change-Id: I8067bef4459ac5b26e250ef046936d487781ee78
Signed-off-by: Fritz Koenig <frkoenig@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2216935
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>