FROMLIST: Add a "nosymfollow" mount option.
For mounts that have the new "nosymfollow" option, don't follow symlinks
when resolving paths. The new option is similar in spirit to the
existing "nodev", "noexec", and "nosuid" options, as well as to the
LOOKUP_NO_SYMLINKS resolve flag in the openat2(2) syscall. Various BSD
variants have been supporting the "nosymfollow" mount option for a long
time with equivalent implementations.

Note that symlinks may still be created on file systems mounted with
the "nosymfollow" option present. readlink() remains functional, so
user space code that is aware of symlinks can still choose to follow
them explicitly.

Setting the "nosymfollow" mount option helps prevent privileged
writers from modifying files unintentionally in case there is an
unexpected link along the accessed path. The "nosymfollow" option is
thus useful as a defensive measure for systems that need to deal with
untrusted file systems in privileged contexts.

More information on the history and motivation for this patch can be
found here:

https://sites.google.com/a/chromium.org/dev/chromium-os/chromiumos-design-docs/hardening-against-malicious-stateful-data#TOC-Restricting-symlink-traversal

Signed-off-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Mattias Nissler <mnissler@chromium.org>
(am from https://patchwork.kernel.org/patch/11420533/)

Conflicts:
	fs/namei.c: Removed check for LOOKUP_NO_SYMLINKS that was added
	in 5.6.

Link: https://lore.kernel.org/linux-fsdevel/20200304173446.122990-1-zwisler@google.com/
BUG=b:152074038,b:146991686,b:152076835
TEST=mount -onosymfollow and try to follow a symlink
Change-Id: Ib2eee4d265018fce162253a2b109d13dd4b33021
Signed-off-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2121041
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Ross Zwisler <zwisler@google.com>