CHROMIUM: ath10k: Update last_tx_bitrate when fixed rate is used
Set the last_tx_bitrate when a fixed rate with a single rate is used.
In this case there is no rate stats updates from firmware.
The last_tx_bitrate needs be set properly in order to calculate estimated
airtime. Some testcases use fixed and failed due to this reason.

This is the same patch as in:
https://eureka-partner-review.googlesource.com/c/qcom/wireless/ath/+/156161

BUG=b:143962220, b:141618158
TEST=stress test using IXIA

Change-Id: I79eaca47b71584c5236c5c1861c08fc4939378bf
Signed-off-by: Kan Yan <kyan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2090681
Reviewed-by: Julan Hsu <julanhsu@google.com>