CHROMIUM: ath10k: Update last_tx_bitrate when fixed rate is used
Set the last_tx_bitrate when a fixed rate with a single rate is used.
In this case there is no rate stats updates from firmware.
The last_tx_bitrate needs be set properly in order to calculate estimated
airtime. Some testcases use fixed and failed due to this reason.

This is the same patch as in:
https://eureka-partner-review.googlesource.com/c/qcom/wireless/ath/+/156161

BUG=b:143962220, b:141618158
TEST=stress test using IXIA

(cherry picked from commit 537d645369804deeb0b65546ddad1c2db805d930
https://chromium.googlesource.com/chromiumos/third_party/kernel/+/refs/heads/chromeos-4.14-gw)
Change-Id: I79eaca47b71584c5236c5c1861c08fc4939378bf
Signed-off-by: Kan Yan <kyan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2090681
Reviewed-by: Julan Hsu <julanhsu@google.com>
Signed-off-by: vamsi singamsetty <vamssi@codeaurora.org>

Change-Id: Ib14af1846edea949a8bff950dcc8040b73d932e6